added pylint to project and update code
