+++ adafruit_character_lcd/character_lcd.py
@@ -1,3 +1,4 @@
+"""Character_LCD - module for interfacing with character lcds
 # The MIT License (MIT)
 #
 # Copyright (c) 2017 Brent Rubell for Adafruit Industries
@@ -19,29 +20,22 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-"""
 `adafruit_CircuitPython_CharLCD`
 ====================================================
-
-
 * Author(s):
 -Brent Rubell
 -Asher Lieber
 -Tony DiCola for the original python charLCD library
-"""
 
-"""
 :mod:`adafruit_character_lcd`
 =================================================
-module for interfacing with character lcds
-"""
+module for interfacing with character lcds"""
+
 import time
-import math
-
 import digitalio
-from board import *
+from micropython import const
 
-
+#pylint: disable-msg=bad-whitespace
 # Commands
 LCD_CLEARDISPLAY        = const(0x01)
 LCD_RETURNHOME          = const(0x02)
@@ -77,8 +71,8 @@ LCD_8BITMODE            = const(0x10)
 LCD_4BITMODE            = const(0x00)
 LCD_2LINE               = const(0x08)
 LCD_1LINE               = const(0x00)
-LCD_5x10DOTS            = const(0x04)
-LCD_5x8DOTS             = const(0x00)
+LCD_5X10DOTS            = const(0x04)
+LCD_5X8DOTS             = const(0x00)
 
 # Offset for up to 4 rows.
 LCD_ROW_OFFSETS         = (0x00, 0x40, 0x14, 0x54)
@@ -101,16 +95,20 @@ _74LS595_LCD_D6          = const(4)
 _74LS595_LCD_D7          = const(3)
 _74LS595_LCD_BACKLIGHT   = const(7)
 
+#pylint: enable-msg=bad-whitespace
 
 def _set_bit(byte_value, position, val):
-    # Given the specified byte_value set the bit at position to the provided
-    # boolean value val and return the modified byte.
+    """Given the specified byte_value set the bit at position to the provided
+    boolean value val and return the modified byte."""
+    ret = None
     if val:
-        return byte_value | (1 << position)
+        ret = byte_value | (1 << position)
     else:
-        return byte_value & ~(1 << position)
+        ret = byte_value & ~(1 << position)
 
+    return ret
 
+#pylint: disable-msg=too-many-instance-attributes
 class Character_LCD(object):
     """ Interfaces with a character LCD
           :param ~digitalio.DigitalInOut rs: The reset data line
@@ -121,179 +119,182 @@ class Character_LCD(object):
           :param ~digitalio.DigitalInOut d7: The data line 7
           :param cols: The columns on the charLCD
           :param lines: The lines on the charLCD
-          :param ~digitalio.DigitalInOut backlight: The backlight pin, usually the last pin. Check with your datasheet
+          :param ~digitalio.DigitalInOut backlight: The backlight pin, usually
+            the last pin. Check with your datasheet
       """
+    #pylint: disable-msg=too-many-arguments
     def __init__(self, rs, en, d4, d5, d6, d7, cols, lines,
-          backlight = None #,
-          #enable_pwm = False,
-          #initial_backlight = 1.0
-          ):
+                 backlight=None #,
+                 #enable_pwm = False,
+                 #initial_backlight = 1.0
+                ):
 
-      self.cols = cols
-      self.lines = lines
-      #  save pin numbers
-      self.rs = rs
-      self.en = en
-      self.d4 = d4
-      self.d5 = d5
-      self.d6 = d6
-      self.d7 = d7
-      # backlight pin
-      self.backlight = backlight
-      # self.pwn_enabled = enable_pwm
-      # set all pins as outputs
-      for pin in(rs, en, d4, d5, d6, d7):
-        pin.direction = digitalio.Direction.OUTPUT
-      #  Setup backlight
-      if backlight is not None:
-        self.backlight.direction = digitalio.Direction.OUTPUT
-        self.backlight.value = 0 # turn backlight on
-      #  initialize the display
-      self._write8(0x33)
-      self._write8(0x32)
-      #  init. display control
-      self.displaycontrol = LCD_DISPLAYON | LCD_CURSOROFF | LCD_BLINKOFF
-      #  init display function
-      self.displayfunction = LCD_4BITMODE | LCD_1LINE | LCD_2LINE | LCD_5x8DOTS
-      #  init display mode
-      self.displaymode = LCD_ENTRYLEFT | LCD_ENTRYSHIFTDECREMENT
-      #  write to display control
-      self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
-      #  write displayfunction
-      self._write8(LCD_FUNCTIONSET | self.displayfunction)
-      #  set the entry mode
-      self._write8(LCD_ENTRYMODESET | self.displaymode)
-      self.clear()
+        self.cols = cols
+        self.lines = lines
+        #  save pin numbers
+        self.reset = rs
+        self.enable = en
+        self.dl4 = d4
+        self.dl5 = d5
+        self.dl6 = d6
+        self.dl7 = d7
+        # backlight pin
+        self.backlight = backlight
+        # self.pwn_enabled = enable_pwm
+        # set all pins as outputs
+        for pin in(rs, en, d4, d5, d6, d7):
+            pin.direction = digitalio.Direction.OUTPUT
+        #  Setup backlight
+        if backlight is not None:
+            self.backlight.direction = digitalio.Direction.OUTPUT
+            self.backlight.value = 0 # turn backlight on
+        #  initialize the display
+        self._write8(0x33)
+        self._write8(0x32)
+        #  init. display control
+        self.displaycontrol = LCD_DISPLAYON | LCD_CURSOROFF | LCD_BLINKOFF
+        #  init display function
+        self.displayfunction = LCD_4BITMODE | LCD_1LINE | LCD_2LINE | LCD_5X8DOTS
+        #  init display mode
+        self.displaymode = LCD_ENTRYLEFT | LCD_ENTRYSHIFTDECREMENT
+        #  write to display control
+        self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
+        #  write displayfunction
+        self._write8(LCD_FUNCTIONSET | self.displayfunction)
+        #  set the entry mode
+        self._write8(LCD_ENTRYMODESET | self.displaymode)
+        self.clear()
+    #pylint: enable-msg=too-many-arguments
 
     def home(self):
-      """Moves the cursor back home pos(1,1)"""
-      self._write8(LCD_RETURNHOME)
-      time.sleep(0.003)
+        """Moves the cursor back home pos(1,1)"""
+        self._write8(LCD_RETURNHOME)
+        time.sleep(0.003)
 
     def clear(self):
-      """Clears the LCD"""
-      self._write8(LCD_CLEARDISPLAY)
-      time.sleep(0.003)
+        """Clears the LCD"""
+        self._write8(LCD_CLEARDISPLAY)
+        time.sleep(0.003)
 
     def show_cursor(self, show):
-      """Show or hide the cursor
+        """Show or hide the cursor
             :param show: True to show cursor, False to hide
-      """
-      if show:
-        self.displaycontrol |= LCD_CURSORON
-      else:
-        self.displaycontrol &= ~LCD_DISPLAYON
-      self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
+        """
+        if show:
+            self.displaycontrol |= LCD_CURSORON
+        else:
+            self.displaycontrol &= ~LCD_DISPLAYON
+        self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
 
     def set_cursor(self, col, row):
-      """Sets the cursor to ``row`` and ``col``
+        """Sets the cursor to ``row`` and ``col``
             :param col: column location
             :param row: row location
         """
-      # Clamp row to the last row of the display
-      if row > self.lines:
-        row = self.lines - 1
-      # Set location
-      self._write8(LCD_SETDDRAMADDR | (col + LCD_ROW_OFFSETS[row]))
+        # Clamp row to the last row of the display
+        if row > self.lines:
+            row = self.lines - 1
+        # Set location
+        self._write8(LCD_SETDDRAMADDR | (col + LCD_ROW_OFFSETS[row]))
 
     def blink(self, blink):
-      """Blinks the cursor if blink = true.
+        """Blinks the cursor if blink = true.
             :param blink: True to blink, False no blink
-      """
-      if blink == True:
-        self.displaycontrol |= LCD_BLINKON
-      else:
-        self.displaycontrol &= ~LCD_BLINKON
-      self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
+        """
+        if blink is True:
+            self.displaycontrol |= LCD_BLINKON
+        else:
+            self.displaycontrol &= ~LCD_BLINKON
+        self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
 
     def move_left(self):
-      """Moves display left one position"""
-      self._write8(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVELEFT)
+        """Moves display left one position"""
+        self._write8(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVELEFT)
 
     def move_right(self):
-      """Moves display right one position"""
-      self._write8(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVERIGHT)
+        """Moves display right one position"""
+        self._write8(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVERIGHT)
 
     def set_left_to_right(self):
-      """Set direction of text to read from left to right"""
-      self.displaymode |= LCD_ENTRYLEFT
-      self._write8(LCD_ENTRYMODESET | self.displaymode)
+        """Set direction of text to read from left to right"""
+        self.displaymode |= LCD_ENTRYLEFT
+        self._write8(LCD_ENTRYMODESET | self.displaymode)
 
     def set_right_to_left(self):
-      """Set direction of text to read from right to left"""
-      self.displaymode |= LCD_ENTRYLEFT
-      self._write8(LCD_ENTRYMODESET | self.displaymode)
+        """Set direction of text to read from right to left"""
+        self.displaymode |= LCD_ENTRYLEFT
+        self._write8(LCD_ENTRYMODESET | self.displaymode)
 
     def enable_display(self, enable):
         """Enable or disable the display.
             :param enable: True to enable display, False to disable
         """
         if enable:
-          self.displaycontrol |= LCD_DISPLAYON
+            self.displaycontrol |= LCD_DISPLAYON
         else:
-          self.displaycontrol &= ~LCD_DISPLAYON
+            self.displaycontrol &= ~LCD_DISPLAYON
         self._write8(LCD_DISPLAYCONTROL | self.displaycontrol)
 
-    def _write8(self,value, char_mode = False):
-      """Sends 8b ``value`` in ``char_mode``.
+    def _write8(self, value, char_mode=False):
+        """Sends 8b ``value`` in ``char_mode``.
         :param value: bytes
         :param char_mode: character/data mode selector. False (default) for
         data only, True for character bits.
         """
-      #  one ms delay to prevent writing too quickly.
-      time.sleep(0.001)
-      #  set character/data bit. (charmode = False)
-      self.rs.value = char_mode
-      # WRITE upper 4 bits
-      self.d4.value = ((value >> 4) & 1) > 0
-      self.d5.value = ((value >> 5) & 1) > 0
-      self.d6.value = ((value >> 6) & 1) > 0
-      self.d7.value = ((value >> 7) & 1) > 0
-      #  send command
-      self._pulse_enable()
-      # WRITE lower 4 bits
-      self.d4.value = (value & 1) > 0
-      self.d5.value = ((value >> 1) & 1) > 0
-      self.d6.value = ((value >> 2) & 1) > 0
-      self.d7.value = ((value >> 3) & 1) > 0
-      self._pulse_enable()
+        #  one ms delay to prevent writing too quickly.
+        time.sleep(0.001)
+        #  set character/data bit. (charmode = False)
+        self.reset.value = char_mode
+        # WRITE upper 4 bits
+        self.dl4.value = ((value >> 4) & 1) > 0
+        self.dl5.value = ((value >> 5) & 1) > 0
+        self.dl6.value = ((value >> 6) & 1) > 0
+        self.dl7.value = ((value >> 7) & 1) > 0
+        #  send command
+        self._pulse_enable()
+        # WRITE lower 4 bits
+        self.dl4.value = (value & 1) > 0
+        self.dl5.value = ((value >> 1) & 1) > 0
+        self.dl6.value = ((value >> 2) & 1) > 0
+        self.dl7.value = ((value >> 3) & 1) > 0
+        self._pulse_enable()
 
     def _pulse_enable(self):
-      """ Pulses (lo->hi->lo) to send commands. """
-      self.en.value = False
-      # 1microsec pause
-      time.sleep(0.0000001)
-      self.en.value = True
-      time.sleep(0.0000001)
-      self.en.value = False
-      time.sleep(0.0000001)
+        """ Pulses (lo->hi->lo) to send commands. """
+        self.enable.value = False
+        # 1microsec pause
+        time.sleep(0.0000001)
+        self.enable.value = True
+        time.sleep(0.0000001)
+        self.enable.value = False
+        time.sleep(0.0000001)
 
     def set_backlight(self, lighton):
-      """ Set lighton to turn the charLCD backlight on.
+        """ Set lighton to turn the charLCD backlight on.
             :param lighton: True to turn backlight on, False to turn off
-      """
-      if lighton:
-        self.backlight.value = 0
-      else:
-        self.backlight.value = 1
+        """
+        if lighton:
+            self.backlight.value = 0
+        else:
+            self.backlight.value = 1
 
 
     def message(self, text):
-      """Write text to display, can include \n for newline
+        """Write text to display, can include \n for newline
             :param text: string to display
-      """
-      line = 0
-      #  iterate thru each char
-      for char in text:
-        # if character is \n, go to next line
-        if char == '\n':
-          line += 1
-          #  move to left/right depending on text direction
-          col = 0 if self.displaymode & LCD_ENTRYLEFT > 0 else self.cols-1
-          self.set_cursor(col, line)
-        # Write character to display
-        else:
-          self._write8(ord(char), True)
+        """
+        line = 0
+        #  iterate thru each char
+        for char in text:
+            # if character is \n, go to next line
+            if char == '\n':
+                line += 1
+                #  move to left/right depending on text direction
+                col = 0 if self.displaymode & LCD_ENTRYLEFT > 0 else self.cols-1
+                self.set_cursor(col, line)
+            # Write character to display
+            else:
+                self._write8(ord(char), True)
 
     def create_char(self, location, pattern):
         """Fill one of the first 8 CGRAM locations with custom characters.
@@ -311,6 +312,7 @@ class Character_LCD(object):
         for i in range(8):
             self._write8(pattern[i], char_mode=True)
 
+#pylint: enable-msg=too-many-instance-attributes
 
 class Character_LCD_I2C(Character_LCD):
     """Character LCD connected to I2C/SPI backpack using its I2C connection.
@@ -329,18 +331,18 @@ class Character_LCD_I2C(Character_LCD):
         # MPY files.  The amount of code and classes implicitly imported
         # by all the SPI and I2C code is too high.  Thus import on demand.
         import adafruit_character_lcd.mcp23008 as mcp23008
-        self._mcp = mcp23008.MCP23008(i2c, address)
+        self._mcp = mcp23008.MCP23008(i2c)
         # Setup pins for I2C backpack, see diagram:
         #   https://learn.adafruit.com/assets/35681
-        rs = self._mcp.DigitalInOut(_MCP23008_LCD_RS, self._mcp)
-        en = self._mcp.DigitalInOut(_MCP23008_LCD_EN, self._mcp)
-        d4 = self._mcp.DigitalInOut(_MCP23008_LCD_D4, self._mcp)
-        d5 = self._mcp.DigitalInOut(_MCP23008_LCD_D5, self._mcp)
-        d6 = self._mcp.DigitalInOut(_MCP23008_LCD_D6, self._mcp)
-        d7 = self._mcp.DigitalInOut(_MCP23008_LCD_D7, self._mcp)
+        reset = self._mcp.DigitalInOut(_MCP23008_LCD_RS, self._mcp)
+        enable = self._mcp.DigitalInOut(_MCP23008_LCD_EN, self._mcp)
+        dl4 = self._mcp.DigitalInOut(_MCP23008_LCD_D4, self._mcp)
+        dl5 = self._mcp.DigitalInOut(_MCP23008_LCD_D5, self._mcp)
+        dl6 = self._mcp.DigitalInOut(_MCP23008_LCD_D6, self._mcp)
+        dl7 = self._mcp.DigitalInOut(_MCP23008_LCD_D7, self._mcp)
         backlight = self._mcp.DigitalInOut(_MCP23008_LCD_BACKLIGHT, self._mcp)
         # Call superclass initializer with MCP23008 pins.
-        super().__init__(rs, en, d4, d5, d6, d7, cols, lines,
+        super().__init__(reset, enable, dl4, dl5, dl6, dl7, cols, lines,
                          backlight=backlight)
 
     def _write8(self, value, char_mode=False):
@@ -388,15 +390,15 @@ class Character_LCD_SPI(Character_LCD):
         self._sr = shift_reg_74ls595.ShiftReg74LS595(spi, latch)
         # Setup pins for SPI backpack, see diagram:
         #   https://learn.adafruit.com/assets/35681
-        rs = self._sr.DigitalInOut(_74LS595_LCD_RS, self._sr)
-        en = self._sr.DigitalInOut(_74LS595_LCD_EN, self._sr)
-        d4 = self._sr.DigitalInOut(_74LS595_LCD_D4, self._sr)
-        d5 = self._sr.DigitalInOut(_74LS595_LCD_D5, self._sr)
-        d6 = self._sr.DigitalInOut(_74LS595_LCD_D6, self._sr)
-        d7 = self._sr.DigitalInOut(_74LS595_LCD_D7, self._sr)
+        reset = self._sr.DigitalInOut(_74LS595_LCD_RS, self._sr)
+        enable = self._sr.DigitalInOut(_74LS595_LCD_EN, self._sr)
+        dl4 = self._sr.DigitalInOut(_74LS595_LCD_D4, self._sr)
+        dl5 = self._sr.DigitalInOut(_74LS595_LCD_D5, self._sr)
+        dl6 = self._sr.DigitalInOut(_74LS595_LCD_D6, self._sr)
+        dl7 = self._sr.DigitalInOut(_74LS595_LCD_D7, self._sr)
         backlight = self._sr.DigitalInOut(_74LS595_LCD_BACKLIGHT, self._sr)
         # Call superclass initializer with shift register pins.
-        super().__init__(rs, en, d4, d5, d6, d7, cols, lines,
+        super().__init__(reset, enable, dl4, dl5, dl6, dl7, cols, lines,
                          backlight=backlight)
 
     def _write8(self, value, char_mode=False):
